Start on refactor
