+++ Cargo.toml
@@ -6,6 +6,7 @@ members = [
   "bench_util",
   "cli",
   "cli/lib",
+  "cli/rt",
   "ext/broadcast_channel",
   "ext/cache",
   "ext/canvas",
@@ -153,6 +154,7 @@ ipnet = "2.3"
 jsonc-parser = { version = "=0.26.2", features = ["serde"] }
 lazy-regex = "3"
 libc = "0.2.168"
+libsui = "0.5.0"
 libz-sys = { version = "1.1.20", default-features = false }
 log = { version = "0.4.20", features = ["kv"] }
 lsp-types = "=0.97.0" # used by tower-lsp and "proposed" feature is unstable in patch releases
