+++ cli/args/mod.rs
@@ -2061,13 +2061,6 @@ pub enum NpmCachingStrategy {
   Manual,
 }
 
-pub fn otel_runtime_config() -> OtelRuntimeConfig {
-  OtelRuntimeConfig {
-    runtime_name: Cow::Borrowed("deno"),
-    runtime_version: Cow::Borrowed(crate::version::DENO_VERSION_INFO.deno),
-  }
-}
-
 #[cfg(test)]
 mod test {
   use pretty_assertions::assert_eq;
