+++ cli/lib/env.rs
@@ -1,5 +1,9 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
+use std::borrow::Cow;
+
+use deno_runtime::deno_telemetry::OtelRuntimeConfig;
+
 pub fn has_trace_permissions_enabled() -> bool {
   has_flag_env_var("DENO_TRACE_PERMISSIONS")
 }
