+++ cli/lib/standalone/mod.rs
@@ -1,3 +1,4 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
+pub mod binary;
 pub mod virtual_fs;
