+++ cli/lib/util/mod.rs
@@ -1,3 +1,7 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
 pub mod checksum;
+pub mod logger;
+pub mod result;
+pub mod text_encoding;
+pub mod v8;
