+++ cli/rt/Cargo.toml
@@ -1,52 +1,18 @@
 # Copyright 2018-2025 the Deno authors. MIT license.
 
 [package]
-name = "deno"
+name = "denort"
 version = "2.1.5"
 authors.workspace = true
-default-run = "deno"
 edition.workspace = true
-exclude = ["bench/testdata/lsp_benchdata/"]
 license.workspace = true
 repository.workspace = true
-description = "Provides the deno executable"
+description = "Provides the denort executable"
 
 [[bin]]
-name = "deno"
+name = "denort"
 path = "main.rs"
 doc = false
-
-[[test]]
-name = "integration"
-path = "integration_tests_runner.rs"
-harness = false
-
-[[bench]]
-name = "deno_bench"
-harness = false
-path = "./bench/main.rs"
-
-[[bench]]
-name = "lsp_bench_standalone"
-harness = false
-path = "./bench/lsp_bench_standalone.rs"
-
-[features]
-default = ["upgrade", "__vendored_zlib_ng"]
-# A feature that enables heap profiling with dhat on Linux.
-# 1. Compile with `cargo build --profile=release-with-debug --features=dhat-heap`
-# 2. Run the executable. It will output a dhat-heap.json file.
-# 3. Open the json file in https://nnethercote.github.io/dh_view/dh_view.html
-dhat-heap = ["dhat"]
-# A feature that enables the upgrade subcommand and the background check for
-# available updates (of deno binary). This is typically disabled for (Linux)
-# distribution packages.
-upgrade = []
-# A dev feature to disable creations and loading of snapshots in favor of
-# loading JS sources at runtime.
-hmr = ["deno_runtime/hmr"]
-# Vendor zlib as zlib-ng
-__vendored_zlib_ng = ["flate2/zlib-ng-compat", "libz-sys/zlib-ng"]
 
 [build-dependencies]
 deno_runtime = { workspace = true, features = ["include_js_files_for_snapshotting", "only_snapshotted_js_sources"] }
@@ -84,7 +50,7 @@ deno_semver.workspace = true
 deno_task_shell = "=0.20.2"
 deno_telemetry.workspace = true
 deno_terminal.workspace = true
-libsui.workspace = true
+libsui = "0.5.0"
 node_resolver.workspace = true
 
 anstream = "0.6.14"
@@ -103,6 +69,7 @@ color-print.workspace = true
 console_static_text.workspace = true
 dashmap.workspace = true
 data-encoding.workspace = true
+deno_media_type.workspace = true
 dhat = { version = "0.3.3", optional = true }
 dissimilar = "=1.0.4"
 dotenvy = "0.15.7"
@@ -110,6 +77,7 @@ dprint-plugin-json = "=0.19.4"
 dprint-plugin-jupyter = "=0.1.5"
 dprint-plugin-markdown = "=0.17.8"
 dprint-plugin-typescript = "=0.93.3"
+env_logger = "=0.10.0"
 fancy-regex = "=0.10.0"
 faster-hex.workspace = true
 # If you disable the default __vendored_zlib_ng feature above, you _must_ be able to link against `-lz`.
@@ -181,14 +149,6 @@ winapi = { workspace = true, features = ["knownfolders", "mswsock", "objbase", "
 nix.workspace = true
 
 [dev-dependencies]
-deno_bench_util.workspace = true
 pretty_assertions.workspace = true
 sys_traits = { workspace = true, features = ["memory"] }
 test_util.workspace = true
-
-[package.metadata.winres]
-# This section defines the metadata that appears in the deno.exe PE header.
-OriginalFilename = "deno.exe"
-LegalCopyright = "© Deno contributors & Deno Land Inc. MIT licensed."
-ProductName = "Deno"
-FileDescription = "Deno: A secure runtime for JavaScript and TypeScript"
