+++ cli/standalone/binary.rs
@@ -72,7 +72,6 @@ use serde::Serialize;
 use super::file_system::DenoCompileFileSystem;
 use super::serialization::deserialize_binary_data_section;
 use super::serialization::serialize_binary_data_section;
-use super::serialization::DenoCompileModuleData;
 use super::serialization::DeserializedDataSection;
 use super::serialization::RemoteModulesStore;
 use super::serialization::RemoteModulesStoreBuilder;
@@ -150,62 +149,6 @@ impl<'a> StandaloneRelativeFileBaseUrl<'a> {
   }
 }
 
-#[derive(Deserialize, Serialize)]
-pub enum NodeModules {
-  Managed {
-    /// Relative path for the node_modules directory in the vfs.
-    node_modules_dir: Option<String>,
-  },
-  Byonm {
-    root_node_modules_dir: Option<String>,
-  },
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct SerializedWorkspaceResolverImportMap {
-  pub specifier: String,
-  pub json: String,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-pub struct SerializedResolverWorkspaceJsrPackage {
-  pub relative_base: String,
-  pub name: String,
-  pub version: Option<Version>,
-  pub exports: IndexMap<String, String>,
-}
-
-#[derive(Deserialize, Serialize)]
-pub struct SerializedWorkspaceResolver {
-  pub import_map: Option<SerializedWorkspaceResolverImportMap>,
-  pub jsr_pkgs: Vec<SerializedResolverWorkspaceJsrPackage>,
-  pub package_jsons: BTreeMap<String, serde_json::Value>,
-  pub pkg_json_resolution: PackageJsonDepResolution,
-}
-
-// Note: Don't use hashmaps/hashsets. Ensure the serialization
-// is deterministic.
-#[derive(Deserialize, Serialize)]
-pub struct Metadata {
-  pub argv: Vec<String>,
-  pub seed: Option<u64>,
-  pub code_cache_key: Option<u64>,
-  pub permissions: PermissionsOptions,
-  pub location: Option<Url>,
-  pub v8_flags: Vec<String>,
-  pub log_level: Option<Level>,
-  pub ca_stores: Option<Vec<String>>,
-  pub ca_data: Option<Vec<u8>>,
-  pub unsafely_ignore_certificate_errors: Option<Vec<String>>,
-  pub env_vars_from_env_file: IndexMap<String, String>,
-  pub workspace_resolver: SerializedWorkspaceResolver,
-  pub entrypoint_key: String,
-  pub node_modules: Option<NodeModules>,
-  pub unstable_config: UnstableConfig,
-  pub otel_config: OtelConfig,
-  pub vfs_case_sensitivity: FileSystemCaseSensitivity,
-}
-
 #[allow(clippy::too_many_arguments)]
 fn write_binary_bytes(
   mut file_writer: File,
@@ -258,146 +201,6 @@ pub fn is_standalone_binary(exe_path: &Path) -> bool {
   libsui::utils::is_elf(&data)
     || libsui::utils::is_pe(&data)
     || libsui::utils::is_macho(&data)
-}
-
-pub struct StandaloneData {
-  pub metadata: Metadata,
-  pub modules: StandaloneModules,
-  pub npm_snapshot: Option<ValidSerializedNpmResolutionSnapshot>,
-  pub root_path: PathBuf,
-  pub source_maps: SourceMapStore,
-  pub vfs: Arc<FileBackedVfs>,
-}
-
-pub struct StandaloneModules {
-  remote_modules: RemoteModulesStore,
-  vfs: Arc<FileBackedVfs>,
-}
-
-impl StandaloneModules {
-  pub fn resolve_specifier<'a>(
-    &'a self,
-    specifier: &'a ModuleSpecifier,
-  ) -> Result<Option<&'a ModuleSpecifier>, AnyError> {
-    if specifier.scheme() == "file" {
-      Ok(Some(specifier))
-    } else {
-      self.remote_modules.resolve_specifier(specifier)
-    }
-  }
-
-  pub fn has_file(&self, path: &Path) -> bool {
-    self.vfs.file_entry(path).is_ok()
-  }
-
-  pub fn read<'a>(
-    &'a self,
-    specifier: &'a ModuleSpecifier,
-    kind: VfsFileSubDataKind,
-  ) -> Result<Option<DenoCompileModuleData<'a>>, AnyError> {
-    if specifier.scheme() == "file" {
-      let path = deno_path_util::url_to_file_path(specifier)?;
-      let bytes = match self.vfs.file_entry(&path) {
-        Ok(entry) => self.vfs.read_file_all(entry, kind)?,
-        Err(err) if err.kind() == ErrorKind::NotFound => {
-          match RealFs.read_file_sync(&path, None) {
-            Ok(bytes) => bytes,
-            Err(FsError::Io(err)) if err.kind() == ErrorKind::NotFound => {
-              return Ok(None)
-            }
-            Err(err) => return Err(err.into()),
-          }
-        }
-        Err(err) => return Err(err.into()),
-      };
-      Ok(Some(DenoCompileModuleData {
-        media_type: MediaType::from_specifier(specifier),
-        specifier,
-        data: bytes,
-      }))
-    } else {
-      self.remote_modules.read(specifier).map(|maybe_entry| {
-        maybe_entry.map(|entry| DenoCompileModuleData {
-          media_type: entry.media_type,
-          specifier: entry.specifier,
-          data: match kind {
-            VfsFileSubDataKind::Raw => entry.data,
-            VfsFileSubDataKind::ModuleGraph => {
-              entry.transpiled_data.unwrap_or(entry.data)
-            }
-          },
-        })
-      })
-    }
-  }
-}
-
-/// This function will try to run this binary as a standalone binary
-/// produced by `deno compile`. It determines if this is a standalone
-/// binary by skipping over the trailer width at the end of the file,
-/// then checking for the magic trailer string `d3n0l4nd`. If found,
-/// the bundle is executed. If not, this function exits with `Ok(None)`.
-pub fn extract_standalone(
-  cli_args: Cow<Vec<OsString>>,
-) -> Result<Option<StandaloneData>, AnyError> {
-  let Some(data) = libsui::find_section("d3n0l4nd") else {
-    return Ok(None);
-  };
-
-  let DeserializedDataSection {
-    mut metadata,
-    npm_snapshot,
-    remote_modules,
-    source_maps,
-    vfs_root_entries,
-    vfs_files_data,
-  } = match deserialize_binary_data_section(data)? {
-    Some(data_section) => data_section,
-    None => return Ok(None),
-  };
-
-  let root_path = {
-    let maybe_current_exe = std::env::current_exe().ok();
-    let current_exe_name = maybe_current_exe
-      .as_ref()
-      .and_then(|p| p.file_name())
-      .map(|p| p.to_string_lossy())
-      // should never happen
-      .unwrap_or_else(|| Cow::Borrowed("binary"));
-    std::env::temp_dir().join(format!("deno-compile-{}", current_exe_name))
-  };
-  let cli_args = cli_args.into_owned();
-  metadata.argv.reserve(cli_args.len() - 1);
-  for arg in cli_args.into_iter().skip(1) {
-    metadata.argv.push(arg.into_string().unwrap());
-  }
-  let vfs = {
-    let fs_root = VfsRoot {
-      dir: VirtualDirectory {
-        // align the name of the directory with the root dir
-        name: root_path.file_name().unwrap().to_string_lossy().to_string(),
-        entries: vfs_root_entries,
-      },
-      root_path: root_path.clone(),
-      start_file_offset: 0,
-    };
-    Arc::new(FileBackedVfs::new(
-      Cow::Borrowed(vfs_files_data),
-      fs_root,
-      metadata.vfs_case_sensitivity,
-    ))
-  };
-  Ok(Some(StandaloneData {
-    metadata,
-    modules: StandaloneModules {
-      remote_modules,
-      vfs: vfs.clone(),
-    },
-    npm_snapshot,
-    root_path,
-    source_maps,
-    vfs,
-  }))
 }
 
 pub struct WriteBinOptions<'a> {
