+++ cli/standalone/virtual_fs.rs
@@ -841,568 +841,6 @@ fn vfs_as_display_tree(
   }
 }
 
-#[derive(Debug)]
-pub struct VfsRoot {
-  pub dir: VirtualDirectory,
-  pub root_path: PathBuf,
-  pub start_file_offset: u64,
-}
-
-impl VfsRoot {
-  fn find_entry<'a>(
-    &'a self,
-    path: &Path,
-    case_sensitivity: FileSystemCaseSensitivity,
-  ) -> std::io::Result<(PathBuf, VfsEntryRef<'a>)> {
-    self.find_entry_inner(path, &mut HashSet::new(), case_sensitivity)
-  }
-
-  fn find_entry_inner<'a>(
-    &'a self,
-    path: &Path,
-    seen: &mut HashSet<PathBuf>,
-    case_sensitivity: FileSystemCaseSensitivity,
-  ) -> std::io::Result<(PathBuf, VfsEntryRef<'a>)> {
-    let mut path = Cow::Borrowed(path);
-    loop {
-      let (resolved_path, entry) =
-        self.find_entry_no_follow_inner(&path, seen, case_sensitivity)?;
-      match entry {
-        VfsEntryRef::Symlink(symlink) => {
-          if !seen.insert(path.to_path_buf()) {
-            return Err(std::io::Error::new(
-              std::io::ErrorKind::Other,
-              "circular symlinks",
-            ));
-          }
-          path = Cow::Owned(symlink.resolve_dest_from_root(&self.root_path));
-        }
-        _ => {
-          return Ok((resolved_path, entry));
-        }
-      }
-    }
-  }
-
-  fn find_entry_no_follow(
-    &self,
-    path: &Path,
-    case_sensitivity: FileSystemCaseSensitivity,
-  ) -> std::io::Result<(PathBuf, VfsEntryRef)> {
-    self.find_entry_no_follow_inner(path, &mut HashSet::new(), case_sensitivity)
-  }
-
-  fn find_entry_no_follow_inner<'a>(
-    &'a self,
-    path: &Path,
-    seen: &mut HashSet<PathBuf>,
-    case_sensitivity: FileSystemCaseSensitivity,
-  ) -> std::io::Result<(PathBuf, VfsEntryRef<'a>)> {
-    let relative_path = match path.strip_prefix(&self.root_path) {
-      Ok(p) => p,
-      Err(_) => {
-        return Err(std::io::Error::new(
-          std::io::ErrorKind::NotFound,
-          "path not found",
-        ));
-      }
-    };
-    let mut final_path = self.root_path.clone();
-    let mut current_entry = VfsEntryRef::Dir(&self.dir);
-    for component in relative_path.components() {
-      let component = component.as_os_str();
-      let current_dir = match current_entry {
-        VfsEntryRef::Dir(dir) => {
-          final_path.push(component);
-          dir
-        }
-        VfsEntryRef::Symlink(symlink) => {
-          let dest = symlink.resolve_dest_from_root(&self.root_path);
-          let (resolved_path, entry) =
-            self.find_entry_inner(&dest, seen, case_sensitivity)?;
-          final_path = resolved_path; // overwrite with the new resolved path
-          match entry {
-            VfsEntryRef::Dir(dir) => {
-              final_path.push(component);
-              dir
-            }
-            _ => {
-              return Err(std::io::Error::new(
-                std::io::ErrorKind::NotFound,
-                "path not found",
-              ));
-            }
-          }
-        }
-        _ => {
-          return Err(std::io::Error::new(
-            std::io::ErrorKind::NotFound,
-            "path not found",
-          ));
-        }
-      };
-      let component = component.to_string_lossy();
-      current_entry = current_dir
-        .entries
-        .get_by_name(&component, case_sensitivity)
-        .ok_or_else(|| {
-          std::io::Error::new(std::io::ErrorKind::NotFound, "path not found")
-        })?
-        .as_ref();
-    }
-
-    Ok((final_path, current_entry))
-  }
-}
-
-pub struct FileBackedVfsFile {
-  file: VirtualFile,
-  pos: RefCell<u64>,
-  vfs: Arc<FileBackedVfs>,
-}
-
-impl FileBackedVfsFile {
-  pub fn seek(&self, pos: SeekFrom) -> std::io::Result<u64> {
-    match pos {
-      SeekFrom::Start(pos) => {
-        *self.pos.borrow_mut() = pos;
-        Ok(pos)
-      }
-      SeekFrom::End(offset) => {
-        if offset < 0 && -offset as u64 > self.file.offset.len {
-          let msg = "An attempt was made to move the file pointer before the beginning of the file.";
-          Err(std::io::Error::new(
-            std::io::ErrorKind::PermissionDenied,
-            msg,
-          ))
-        } else {
-          let mut current_pos = self.pos.borrow_mut();
-          *current_pos = if offset >= 0 {
-            self.file.offset.len - (offset as u64)
-          } else {
-            self.file.offset.len + (-offset as u64)
-          };
-          Ok(*current_pos)
-        }
-      }
-      SeekFrom::Current(offset) => {
-        let mut current_pos = self.pos.borrow_mut();
-        if offset >= 0 {
-          *current_pos += offset as u64;
-        } else if -offset as u64 > *current_pos {
-          return Err(std::io::Error::new(std::io::ErrorKind::PermissionDenied, "An attempt was made to move the file pointer before the beginning of the file."));
-        } else {
-          *current_pos -= -offset as u64;
-        }
-        Ok(*current_pos)
-      }
-    }
-  }
-
-  pub fn read_to_buf(&self, buf: &mut [u8]) -> std::io::Result<usize> {
-    let read_pos = {
-      let mut pos = self.pos.borrow_mut();
-      let read_pos = *pos;
-      // advance the position due to the read
-      *pos = std::cmp::min(self.file.offset.len, *pos + buf.len() as u64);
-      read_pos
-    };
-    self.vfs.read_file(&self.file, read_pos, buf)
-  }
-
-  fn read_to_end(&self) -> FsResult<Cow<'static, [u8]>> {
-    let read_pos = {
-      let mut pos = self.pos.borrow_mut();
-      let read_pos = *pos;
-      // todo(dsherret): should this always set it to the end of the file?
-      if *pos < self.file.offset.len {
-        // advance the position due to the read
-        *pos = self.file.offset.len;
-      }
-      read_pos
-    };
-    if read_pos > self.file.offset.len {
-      return Ok(Cow::Borrowed(&[]));
-    }
-    if read_pos == 0 {
-      Ok(
-        self
-          .vfs
-          .read_file_all(&self.file, VfsFileSubDataKind::Raw)?,
-      )
-    } else {
-      let size = (self.file.offset.len - read_pos) as usize;
-      let mut buf = vec![0; size];
-      self.vfs.read_file(&self.file, read_pos, &mut buf)?;
-      Ok(Cow::Owned(buf))
-    }
-  }
-}
-
-#[async_trait::async_trait(?Send)]
-impl deno_io::fs::File for FileBackedVfsFile {
-  fn read_sync(self: Rc<Self>, buf: &mut [u8]) -> FsResult<usize> {
-    self.read_to_buf(buf).map_err(Into::into)
-  }
-  async fn read_byob(
-    self: Rc<Self>,
-    mut buf: BufMutView,
-  ) -> FsResult<(usize, BufMutView)> {
-    // this is fast, no need to spawn a task
-    let nread = self.read_to_buf(&mut buf)?;
-    Ok((nread, buf))
-  }
-
-  fn write_sync(self: Rc<Self>, _buf: &[u8]) -> FsResult<usize> {
-    Err(FsError::NotSupported)
-  }
-  async fn write(
-    self: Rc<Self>,
-    _buf: BufView,
-  ) -> FsResult<deno_core::WriteOutcome> {
-    Err(FsError::NotSupported)
-  }
-
-  fn write_all_sync(self: Rc<Self>, _buf: &[u8]) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn write_all(self: Rc<Self>, _buf: BufView) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  fn read_all_sync(self: Rc<Self>) -> FsResult<Cow<'static, [u8]>> {
-    self.read_to_end()
-  }
-  async fn read_all_async(self: Rc<Self>) -> FsResult<Cow<'static, [u8]>> {
-    // this is fast, no need to spawn a task
-    self.read_to_end()
-  }
-
-  fn chmod_sync(self: Rc<Self>, _pathmode: u32) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn chmod_async(self: Rc<Self>, _mode: u32) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  fn seek_sync(self: Rc<Self>, pos: SeekFrom) -> FsResult<u64> {
-    self.seek(pos).map_err(|err| err.into())
-  }
-  async fn seek_async(self: Rc<Self>, pos: SeekFrom) -> FsResult<u64> {
-    self.seek(pos).map_err(|err| err.into())
-  }
-
-  fn datasync_sync(self: Rc<Self>) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn datasync_async(self: Rc<Self>) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  fn sync_sync(self: Rc<Self>) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn sync_async(self: Rc<Self>) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  fn stat_sync(self: Rc<Self>) -> FsResult<FsStat> {
-    Err(FsError::NotSupported)
-  }
-  async fn stat_async(self: Rc<Self>) -> FsResult<FsStat> {
-    Err(FsError::NotSupported)
-  }
-
-  fn lock_sync(self: Rc<Self>, _exclusive: bool) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn lock_async(self: Rc<Self>, _exclusive: bool) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  fn unlock_sync(self: Rc<Self>) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn unlock_async(self: Rc<Self>) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  fn truncate_sync(self: Rc<Self>, _len: u64) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn truncate_async(self: Rc<Self>, _len: u64) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  fn utime_sync(
-    self: Rc<Self>,
-    _atime_secs: i64,
-    _atime_nanos: u32,
-    _mtime_secs: i64,
-    _mtime_nanos: u32,
-  ) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-  async fn utime_async(
-    self: Rc<Self>,
-    _atime_secs: i64,
-    _atime_nanos: u32,
-    _mtime_secs: i64,
-    _mtime_nanos: u32,
-  ) -> FsResult<()> {
-    Err(FsError::NotSupported)
-  }
-
-  // lower level functionality
-  fn as_stdio(self: Rc<Self>) -> FsResult<std::process::Stdio> {
-    Err(FsError::NotSupported)
-  }
-  fn backing_fd(self: Rc<Self>) -> Option<ResourceHandleFd> {
-    None
-  }
-  fn try_clone_inner(self: Rc<Self>) -> FsResult<Rc<dyn deno_io::fs::File>> {
-    Ok(self)
-  }
-}
-
-#[derive(Debug, Clone)]
-pub struct FileBackedVfsDirEntry {
-  pub parent_path: PathBuf,
-  pub metadata: FileBackedVfsMetadata,
-}
-
-#[derive(Debug, Clone)]
-pub struct FileBackedVfsMetadata {
-  pub name: String,
-  pub file_type: sys_traits::FileType,
-  pub len: u64,
-}
-
-impl FileBackedVfsMetadata {
-  pub fn from_vfs_entry_ref(vfs_entry: VfsEntryRef) -> Self {
-    FileBackedVfsMetadata {
-      file_type: match vfs_entry {
-        VfsEntryRef::Dir(_) => sys_traits::FileType::Dir,
-        VfsEntryRef::File(_) => sys_traits::FileType::File,
-        VfsEntryRef::Symlink(_) => sys_traits::FileType::Symlink,
-      },
-      name: vfs_entry.name().to_string(),
-      len: match vfs_entry {
-        VfsEntryRef::Dir(_) => 0,
-        VfsEntryRef::File(file) => file.offset.len,
-        VfsEntryRef::Symlink(_) => 0,
-      },
-    }
-  }
-  pub fn as_fs_stat(&self) -> FsStat {
-    FsStat {
-      is_directory: self.file_type == sys_traits::FileType::Dir,
-      is_file: self.file_type == sys_traits::FileType::File,
-      is_symlink: self.file_type == sys_traits::FileType::Symlink,
-      atime: None,
-      birthtime: None,
-      mtime: None,
-      ctime: None,
-      blksize: 0,
-      size: self.len,
-      dev: 0,
-      ino: 0,
-      mode: 0,
-      nlink: 0,
-      uid: 0,
-      gid: 0,
-      rdev: 0,
-      blocks: 0,
-      is_block_device: false,
-      is_char_device: false,
-      is_fifo: false,
-      is_socket: false,
-    }
-  }
-}
-
-#[derive(Debug)]
-pub struct FileBackedVfs {
-  vfs_data: Cow<'static, [u8]>,
-  fs_root: VfsRoot,
-  case_sensitivity: FileSystemCaseSensitivity,
-}
-
-impl FileBackedVfs {
-  pub fn new(
-    data: Cow<'static, [u8]>,
-    fs_root: VfsRoot,
-    case_sensitivity: FileSystemCaseSensitivity,
-  ) -> Self {
-    Self {
-      vfs_data: data,
-      fs_root,
-      case_sensitivity,
-    }
-  }
-
-  pub fn root(&self) -> &Path {
-    &self.fs_root.root_path
-  }
-
-  pub fn is_path_within(&self, path: &Path) -> bool {
-    path.starts_with(&self.fs_root.root_path)
-  }
-
-  pub fn open_file(
-    self: &Arc<Self>,
-    path: &Path,
-  ) -> std::io::Result<FileBackedVfsFile> {
-    let file = self.file_entry(path)?;
-    Ok(FileBackedVfsFile {
-      file: file.clone(),
-      vfs: self.clone(),
-      pos: Default::default(),
-    })
-  }
-
-  pub fn read_dir(&self, path: &Path) -> std::io::Result<Vec<FsDirEntry>> {
-    let dir = self.dir_entry(path)?;
-    Ok(
-      dir
-        .entries
-        .iter()
-        .map(|entry| FsDirEntry {
-          name: entry.name().to_string(),
-          is_file: matches!(entry, VfsEntry::File(_)),
-          is_directory: matches!(entry, VfsEntry::Dir(_)),
-          is_symlink: matches!(entry, VfsEntry::Symlink(_)),
-        })
-        .collect(),
-    )
-  }
-
-  pub fn read_dir_with_metadata<'a>(
-    &'a self,
-    path: &Path,
-  ) -> std::io::Result<impl Iterator<Item = FileBackedVfsDirEntry> + 'a> {
-    let dir = self.dir_entry(path)?;
-    let path = path.to_path_buf();
-    Ok(dir.entries.iter().map(move |entry| FileBackedVfsDirEntry {
-      parent_path: path.to_path_buf(),
-      metadata: FileBackedVfsMetadata::from_vfs_entry_ref(entry.as_ref()),
-    }))
-  }
-
-  pub fn read_link(&self, path: &Path) -> std::io::Result<PathBuf> {
-    let (_, entry) = self
-      .fs_root
-      .find_entry_no_follow(path, self.case_sensitivity)?;
-    match entry {
-      VfsEntryRef::Symlink(symlink) => {
-        Ok(symlink.resolve_dest_from_root(&self.fs_root.root_path))
-      }
-      VfsEntryRef::Dir(_) | VfsEntryRef::File(_) => Err(std::io::Error::new(
-        std::io::ErrorKind::Other,
-        "not a symlink",
-      )),
-    }
-  }
-
-  pub fn lstat(&self, path: &Path) -> std::io::Result<FileBackedVfsMetadata> {
-    let (_, entry) = self
-      .fs_root
-      .find_entry_no_follow(path, self.case_sensitivity)?;
-    Ok(FileBackedVfsMetadata::from_vfs_entry_ref(entry))
-  }
-
-  pub fn stat(&self, path: &Path) -> std::io::Result<FileBackedVfsMetadata> {
-    let (_, entry) = self.fs_root.find_entry(path, self.case_sensitivity)?;
-    Ok(FileBackedVfsMetadata::from_vfs_entry_ref(entry))
-  }
-
-  pub fn canonicalize(&self, path: &Path) -> std::io::Result<PathBuf> {
-    let (path, _) = self.fs_root.find_entry(path, self.case_sensitivity)?;
-    Ok(path)
-  }
-
-  pub fn read_file_all(
-    &self,
-    file: &VirtualFile,
-    sub_data_kind: VfsFileSubDataKind,
-  ) -> std::io::Result<Cow<'static, [u8]>> {
-    let read_len = match sub_data_kind {
-      VfsFileSubDataKind::Raw => file.offset.len,
-      VfsFileSubDataKind::ModuleGraph => file.module_graph_offset.len,
-    };
-    let read_range = self.get_read_range(file, sub_data_kind, 0, read_len)?;
-    match &self.vfs_data {
-      Cow::Borrowed(data) => Ok(Cow::Borrowed(&data[read_range])),
-      Cow::Owned(data) => Ok(Cow::Owned(data[read_range].to_vec())),
-    }
-  }
-
-  pub fn read_file(
-    &self,
-    file: &VirtualFile,
-    pos: u64,
-    buf: &mut [u8],
-  ) -> std::io::Result<usize> {
-    let read_range = self.get_read_range(
-      file,
-      VfsFileSubDataKind::Raw,
-      pos,
-      buf.len() as u64,
-    )?;
-    let read_len = read_range.len();
-    buf[..read_len].copy_from_slice(&self.vfs_data[read_range]);
-    Ok(read_len)
-  }
-
-  fn get_read_range(
-    &self,
-    file: &VirtualFile,
-    sub_data_kind: VfsFileSubDataKind,
-    pos: u64,
-    len: u64,
-  ) -> std::io::Result<Range<usize>> {
-    let file_offset_and_len = match sub_data_kind {
-      VfsFileSubDataKind::Raw => file.offset,
-      VfsFileSubDataKind::ModuleGraph => file.module_graph_offset,
-    };
-    if pos > file_offset_and_len.len {
-      return Err(std::io::Error::new(
-        std::io::ErrorKind::UnexpectedEof,
-        "unexpected EOF",
-      ));
-    }
-    let file_offset =
-      self.fs_root.start_file_offset + file_offset_and_len.offset;
-    let start = file_offset + pos;
-    let end = file_offset + std::cmp::min(pos + len, file_offset_and_len.len);
-    Ok(start as usize..end as usize)
-  }
-
-  pub fn dir_entry(&self, path: &Path) -> std::io::Result<&VirtualDirectory> {
-    let (_, entry) = self.fs_root.find_entry(path, self.case_sensitivity)?;
-    match entry {
-      VfsEntryRef::Dir(dir) => Ok(dir),
-      VfsEntryRef::Symlink(_) => unreachable!(),
-      VfsEntryRef::File(_) => Err(std::io::Error::new(
-        std::io::ErrorKind::Other,
-        "path is a file",
-      )),
-    }
-  }
-
-  pub fn file_entry(&self, path: &Path) -> std::io::Result<&VirtualFile> {
-    let (_, entry) = self.fs_root.find_entry(path, self.case_sensitivity)?;
-    match entry {
-      VfsEntryRef::Dir(_) => Err(std::io::Error::new(
-        std::io::ErrorKind::Other,
-        "path is a directory",
-      )),
-      VfsEntryRef::Symlink(_) => unreachable!(),
-      VfsEntryRef::File(file) => Ok(file),
-    }
-  }
-}
-
 #[cfg(test)]
 mod test {
   use std::io::Write;
