+++ cli/util/mod.rs
@@ -9,10 +9,8 @@ pub mod draw_thread;
 pub mod extract;
 pub mod file_watcher;
 pub mod fs;
-pub mod logger;
 pub mod path;
 pub mod progress_bar;
-pub mod result;
 pub mod retry;
 pub mod sync;
 pub mod text_encoding;
