+++ cli/util/text_encoding.rs
@@ -11,24 +11,6 @@ use deno_core::ModuleSourceCode;
 static SOURCE_MAP_PREFIX: &[u8] =
   b"//# sourceMappingURL=data:application/json;base64,";
 
-#[inline(always)]
-pub fn from_utf8_lossy_cow(bytes: Cow<[u8]>) -> Cow<str> {
-  match bytes {
-    Cow::Borrowed(bytes) => String::from_utf8_lossy(bytes),
-    Cow::Owned(bytes) => Cow::Owned(from_utf8_lossy_owned(bytes)),
-  }
-}
-
-#[inline(always)]
-pub fn from_utf8_lossy_owned(bytes: Vec<u8>) -> String {
-  match String::from_utf8_lossy(&bytes) {
-    Cow::Owned(code) => code,
-    // SAFETY: `String::from_utf8_lossy` guarantees that the result is valid
-    // UTF-8 if `Cow::Borrowed` is returned.
-    Cow::Borrowed(_) => unsafe { String::from_utf8_unchecked(bytes) },
-  }
-}
-
 pub fn source_map_from_code(code: &[u8]) -> Option<Vec<u8>> {
   let range = find_source_map_range(code)?;
   let source_map_range = &code[range];
@@ -103,29 +85,6 @@ fn find_source_map_range(code: &[u8]) -> Option<Range<usize>> {
   } else {
     None
   }
-}
-
-/// Converts an `Arc<str>` to an `Arc<[u8]>`.
-#[allow(dead_code)]
-pub fn arc_str_to_bytes(arc_str: Arc<str>) -> Arc<[u8]> {
-  let raw = Arc::into_raw(arc_str);
-  // SAFETY: This is safe because they have the same memory layout.
-  unsafe { Arc::from_raw(raw as *const [u8]) }
-}
-
-/// Converts an `Arc<u8>` to an `Arc<str>` if able.
-#[allow(dead_code)]
-pub fn arc_u8_to_arc_str(
-  arc_u8: Arc<[u8]>,
-) -> Result<Arc<str>, std::str::Utf8Error> {
-  // Check that the string is valid UTF-8.
-  std::str::from_utf8(&arc_u8)?;
-  // SAFETY: the string is valid UTF-8, and the layout Arc<[u8]> is the same as
-  // Arc<str>. This is proven by the From<Arc<str>> impl for Arc<[u8]> from the
-  // standard library.
-  Ok(unsafe {
-    std::mem::transmute::<std::sync::Arc<[u8]>, std::sync::Arc<str>>(arc_u8)
-  })
 }
 
 #[cfg(test)]
